+++ promgen_k8s/pods_job.py
@@ -46,6 +46,7 @@ class PodsJob:
       prom_conf['scrape_configs'][-1]['relabel_configs'].extend([
         replace(source_labels=['__meta_kubernetes_pod_host_ip','__address__','__metrics_path__'],
           regex='(172\\.2[0,2-9].[0-9]*\\.[0-9]*);([0-9]*\\.[0-9]*\\.[0-9]*\\.[0-9]*:[0-9]*);(.*)', replacement='/proxy/$2$3',
+        copy_value('__address__', 'instance'),
           target_label='__metrics_path__'),
         replace(source_labels=['__meta_kubernetes_pod_host_ip'],
           regex='(172\\.2[0,2-9].[0-9]*\\.[0-9]*)', replacement=c.proxy,