+++ promgen_k8s/service_endpoints_job.py
@@ -42,6 +42,7 @@ class ServiceEndpointsJob:
 
     if not c.incluster:
       prom_conf['scrape_configs'][-1]['relabel_configs'].extend([
+        copy_value('__address__', 'instance'),
         replace(source_labels=['__address__', '__metrics_path__'],
           regex='([0-9]*\\.[0-9]*\\.[0-9]*\\.[0-9]*:[0-9]*);(.*)', replacement='/proxy/$1$2',
           target_label='__metrics_path__'),