+++ promgen_k8s/service_endpoints_job.py
@@ -2,8 +2,9 @@
 from prom_dsl import *
 
 class ServiceEndpointsJob:
-  def __init__(self, additional_relabel_configs=[], additional_metric_relabel_configs=[]):
+  def __init__(self, scrape_interval=None, additional_relabel_configs=[], additional_metric_relabel_configs=[]):
     self.type = 'service-endpoints'
+    self.scrape_interval = scrape_interval
     self.additional_relabel_configs = additional_relabel_configs
     self.additional_metric_relabel_configs = additional_metric_relabel_configs
 
@@ -48,6 +49,10 @@ class ServiceEndpointsJob:
           target_label='__metrics_path__'),
         set_value('__address__', c.proxy)
       ])
+
+    # set job's scrape_interval if defined
+    if not self.scrape_interval is None:
+      prom_conf['scrape_configs'][-1]['scrape_interval'] = self.scrape_interval
 
     # add additional configs
     prom_conf['scrape_configs'][-1]['relabel_configs'].extend(self.additional_relabel_configs)