+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 IMPROVEMENTS:
 
 - Remove some unneeded since duplicated labels from pods, nodes and cadvisor jobs
+- Fix typo in name of the `ServiceEndpointsJob`
 
 ## 0.1.2 (November 7, 2017)