+++ CHANGELOG.md
@@ -6,6 +6,7 @@ IMPROVEMENTS:
 
 - Remove some unneeded since duplicated labels from pods, nodes and cadvisor jobs
 - Fix typo in name of the `ServiceEndpointsJob`
+- Provide unique instance label for all jobs, usually copied from the __address__ label
 
 ## 0.1.2 (November 7, 2017)